Update City.java
Replaced flight variables with airport versions and added final string airportName. I'm unsure of what else to do with this at the moment. I'll have to look into this again another time when I have a better idea of what to do with it but I wanted to contribute something at this point, even if it probably still needs plenty of adjustments.